[VXP]: Delete old ui elements.
