+++ VirtualApp/app/src/main/java/io/virtualapp/splash/SplashActivity.java
@@ -10,7 +10,7 @@ import io.virtualapp.VCommends;
 import io.virtualapp.abs.ui.VActivity;
 import io.virtualapp.abs.ui.VUiKit;
 import io.virtualapp.home.FlurryROMCollector;
-import io.virtualapp.home.HomeActivity;
+import io.virtualapp.home.NewHomeActivity;
 import jonathanfinerty.once.Once;
 
 public class SplashActivity extends VActivity {
@@ -37,7 +37,7 @@ public class SplashActivity extends VActivity {
                 VUiKit.sleep(delta);
             }
         }).done((res) -> {
-            HomeActivity.goHome(this);
+            NewHomeActivity.goHome(this);
             finish();
         });
     }
